calparks_project/requirements/base.txt
Django==1.5.1
bpython==0.12
django-braces==1.0.0
django-model-utils==1.3.1
logutils==0.3.3
South==0.7.6
django-registration==0.8
gunicorn==0.17.0
django-pagination==1.0.7
Unipath==0.2.1
MySQL-python==1.2.4c1
django-crispy-forms==1.2.8
django-filter
python-dateutil==1.5
django-compressor
versiontools
django-debug-toolbar
